requirements.txt
pandas
numpy
torch
transformers
scikit-learn
spacy
fastapi
uvicorn
pyyaml
dvc